# docs/relay/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UAM Relay",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/inbox/{address}": {
      "get": {
        "summary": "Get Inbox",
        "description": "Retrieve stored messages for an agent.\n\nBearer token auth via ``verify_token_http`` dependency.\nAgent can only read their own inbox.",
        "operationId": "get_inbox_api_v1_inbox__address__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboxResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/thread/{thread_id}": {
      "get": {
        "summary": "Get Thread Messages",
        "description": "Retrieve messages in a thread.\n\nRequires Bearer token auth.  The authenticated agent must be a\nparticipant (from_addr or to_addr) in at least one message in the\nthread.  Returns 403 if not a participant or if thread is empty.",
        "operationId": "get_thread_messages_api_v1_messages_thread__thread_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/{message_id}/receipt": {
      "post": {
        "summary": "Submit Receipt",
        "description": "Submit a receipt (e.g. ``receipt.read``) for a received message.\n\nRequires Bearer token auth.  The authenticated agent must be the\nrecipient of the message.  Routes the receipt envelope to the\noriginal sender via the connection manager.",
        "operationId": "submit_receipt_api_v1_messages__message_id__receipt_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/{address}/presence": {
      "get": {
        "summary": "Get Presence",
        "description": "Check whether an agent is currently online.\n\nReturns the agent's online status (based on active WebSocket connection)\nand their last_seen timestamp from the database.\n\nRequires Bearer token authentication.",
        "operationId": "get_presence_api_v1_agents__address__presence_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/handshakes/send": {
      "post": {
        "tags": [
          "handshakes"
        ],
        "summary": "Send Handshake",
        "description": "Initiate a relay-mediated handshake with another agent.\n\nCreates a pending handshake request from the authenticated agent\nto the target address.  Optionally includes a contact card.",
        "operationId": "send_handshake_api_v1_handshakes_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandshakeSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandshakeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/handshakes/pending/{address}": {
      "get": {
        "tags": [
          "handshakes"
        ],
        "summary": "List Pending Handshakes",
        "description": "List pending handshakes for the authenticated agent.\n\nAgent can only view their own pending handshakes (incoming).",
        "operationId": "list_pending_handshakes_api_v1_handshakes_pending__address__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandshakeListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/handshakes/{handshake_id}/respond": {
      "post": {
        "tags": [
          "handshakes"
        ],
        "summary": "Respond To Handshake",
        "description": "Respond to a pending handshake (approve or deny).\n\nThe authenticated agent must be the ``to_addr`` of the handshake.",
        "operationId": "respond_to_handshake_api_v1_handshakes__handshake_id__respond_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "handshake_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Handshake Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandshakeRespondRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandshakeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HandshakeListResponse": {
        "properties": {
          "handshakes": {
            "items": {
              "$ref": "#/components/schemas/HandshakeResponse"
            },
            "type": "array",
            "title": "Handshakes"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "handshakes",
          "count"
        ],
        "title": "HandshakeListResponse"
      },
      "HandshakeRespondRequest": {
        "properties": {
          "response": {
            "type": "string",
            "title": "Response"
          }
        },
        "type": "object",
        "required": [
          "response"
        ],
        "title": "HandshakeRespondRequest"
      },
      "HandshakeResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "from_addr": {
            "type": "string",
            "title": "From Addr"
          },
          "to_addr": {
            "type": "string",
            "title": "To Addr"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "from_addr",
          "to_addr"
        ],
        "title": "HandshakeResponse"
      },
      "HandshakeSendRequest": {
        "properties": {
          "to_address": {
            "type": "string",
            "title": "To Address"
          },
          "contact_card": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Card"
          }
        },
        "type": "object",
        "required": [
          "to_address"
        ],
        "title": "HandshakeSendRequest"
      },
...
      "InboxResponse": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          },
          "messages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Messages"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "address",
          "messages",
          "count"
        ],
        "title": "InboxResponse"
      },
      "PresenceResponse": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          },
          "online": {
            "type": "boolean",
            "title": "Online"
          },
          "last_seen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          }
        },
        "type": "object",
        "required": [
          "address",
          "online"
        ],
        "title": "PresenceResponse"
      },
...
      "ReceiptRequest": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "ReceiptRequest"
      },
      "ReceiptResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message_id"
        ],
        "title": "ReceiptResponse"
      },
...
      "ThreadResponse": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "messages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Messages"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "messages",
          "count"
        ],
        "title": "ThreadResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}